Initial cut of moving Game code from Driver to GameEstimator
