+++ photon-ml/src/integTest/scala/com/linkedin/photon/ml/optimization/OptimizerIntegTest.scala
@@ -17,7 +17,6 @@ package com.linkedin.photon.ml.optimization
 import java.util.Random
 
 import breeze.linalg.{DenseVector, SparseVector, norm}
-import breeze.optimize.FirstOrderMinimizer.{FunctionValuesConverged, GradientConverged}
 import org.apache.spark.Logging
 import org.apache.spark.broadcast.Broadcast
 import org.mockito.Mockito._
@@ -28,6 +27,7 @@ import com.linkedin.photon.ml.data.LabeledPoint
 import com.linkedin.photon.ml.function.TwiceDiffFunction
 import com.linkedin.photon.ml.normalization.{NoNormalization, NormalizationContext}
 import com.linkedin.photon.ml.test.SparkTestUtils
+import com.linkedin.photon.ml.util.{FunctionValuesConverged, GradientConverged}
 
 /**
  * Verify that core optimizers do reasonable things on small test problems.
